fix: Updated dropwizard3 to a version with the separated jmx module.
