+++ ratis-metrics/pom.xml
@@ -62,18 +62,11 @@
     </dependency>
     <dependency>
       <groupId>io.dropwizard.metrics</groupId>
-      <artifactId>metrics-jvm</artifactId>
-      <optional>true</optional>
+      <artifactId>metrics-jmx</artifactId>
     </dependency>
     <dependency>
       <groupId>io.dropwizard.metrics</groupId>
-      <artifactId>metrics-ganglia</artifactId>
-      <exclusions>
-        <exclusion>
-          <groupId>org.acplt</groupId>
-          <artifactId>oncrpc</artifactId>
-        </exclusion>
-      </exclusions>
+      <artifactId>metrics-jvm</artifactId>
       <optional>true</optional>
     </dependency>
   </dependencies>
